// layers/ui/app/common/components/ContextMenu/ContextMenu.types.ts
export interface ContextMenuItem {
	/** Текст пункта меню. */
	label: string
	/** Действие при клике */
	onClick?: () => unknown
	/** Вложенные пункты, которые отображаются в подменю. */
	children?: ContextMenuItem[]
	/** Отключает пункт меню и запрещает взаимодействие с ним. */
	disabled?: boolean
}